requirements.txt
python-telegram-bot[job-queue]>=20.7
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.3
tinydb>=4.8.0